Refactor admin's hard_merge code, and only allow one item at a time.

+++ astrobin/admin.py
@@ -137,88 +137,14 @@ class GearAssistedMergeAdmin(admin.ModelAdmin):
     soft_merge.short_description = 'Soft merge'
 
     def hard_merge(modeladmin, request, queryset):
-        import operator
+        from utils import unique_items
+        masters = unique_items([x.master for x in queryset])
+        if len(masters) > 1:
+            return
 
+        master = masters[0]
         for merge in queryset:
-            merge_done = False
-            # Find matching slaves in images
-            types = {
-                'imaging_telescopes': Telescope,
-                'guiding_telescopes': Telescope,
-                'mounts': Mount,
-                'imaging_cameras': Camera,
-                'guiding_cameras': Camera,
-                'focal_reducers': FocalReducer,
-                'software': Software,
-                'filters': Filter,
-                'accessories': Accessory,
-            }
-            filters = reduce(operator.or_, [Q(**{'%s__gear_ptr__pk' % t: merge.slave.pk}) for t in types])
-            images = Image.objects.filter(filters).distinct()
-            for image in images:
-                changed = False
-                for name, klass in types.iteritems():
-                    slave = getattr(image, name).filter(pk = merge.slave.pk)
-                    if slave:
-                        try:
-                            getattr(image, name).add(klass.objects.get(gear_ptr = merge.master))
-                            getattr(image, name).remove(slave[0])
-                            merge_done = True
-                            changed = True
-                        except klass.DoesNotExist:
-                            continue
-                if changed:
-                    image.save()
-
-            # Find matching slaves in user profiles
-            types = {
-                'telescopes': Telescope,
-                'mounts': Mount,
-                'cameras': Camera,
-                'focal_reducers': FocalReducer,
-                'software': Software,
-                'filters': Filter,
-                'accessories': Accessory,
-            }
-            filters = reduce(operator.or_, [Q(**{'%s__gear_ptr__pk' % t: merge.slave.pk}) for t in types])
-            owners = UserProfile.objects.filter(filters).distinct()
-            for owner in owners:
-                changed = False
-                for name, klass in types.iteritems():
-                    slave = getattr(owner, name).filter(pk = merge.slave.pk)
-                    if slave:
-                        try:
-                            getattr(owner, name).add(klass.objects.get(gear_ptr = merge.master))
-                            getattr(owner, name).remove(slave[0])
-                            merge_done = True
-                            changed = True
-                        except klass.DoesNotExist:
-                            continue
-                if changed:
-                    owner.save()
-
-            # Find matching slaves in deep sky acquisitions
-            try:
-                filter = Filter.objects.get(gear_ptr__pk = merge.master.pk)
-                DeepSky_Acquisition.objects.filter(filter__gear_ptr__pk = merge.slave.pk).update(
-                    filter = filter)
-            except Filter.DoesNotExist:
-                pass
-
-            # Find matching gear comments and move them to the master
-            GearComment.objects.filter(gear = merge.slave).update(gear = merge.master)
-
-            # Find matching gear reviews and move them to the master
-            reviews = ReviewedItem.objects.filter(gear = merge.slave).update(object_id = merge.master.id)
-
-            # Fetch slave's master if this hard-merge's master doesn't have a soft-merge master
-            if not merge.master.master:
-                merge.master.master = merge.slave.master
-
-        # Only now, delete all the slaves. We must delete at the end because
-        # We might have the same slave respond to different masters.
-        for merge in queryset:
-            merge.slave.delete()
+            master.hard_merge(merge.slave)
 
         # Finally, clear up the temporary model
         queryset.delete()
